+++ ecommerce-microservices/order-service/src/main/java/com/toleyko/springboot/orderservice/dao/OrderRepository.java
@@ -6,9 +6,8 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
-public interface OrderRepository extends JpaRepository<Order, Integer> {
+public interface OrderRepository extends JpaRepository<Order, Long> {
     public List<Order> getOrderByUsername(String username);
-
     @Transactional
     public void deleteAllByUsername(String username);
 }
