+++ ecommerce-microservices/order-service/src/main/java/com/toleyko/springboot/orderservice/service/OrderService.java
@@ -7,10 +7,10 @@ import java.util.List;
 
 public interface OrderService {
     public List<Order> getAllOrders();
-    public Order getOrderById(Integer id) throws OrderNotFoundException;
-    public Order deleteOrderById(Integer id) throws OrderNotFoundException;
+    public Order getOrderById(Long id) throws OrderNotFoundException;
+    public Order deleteOrderById(Long id) throws OrderNotFoundException;
     public Order saveOrder(Order order);
-    public Order updateOrderById(Integer id, Order order) throws OrderNotFoundException;
+    public Order updateOrderById(Long id, Order order) throws OrderNotFoundException;
     public List<Order> getOrdersByUsername(String username) throws OrderNotFoundException;
     public void deleteOrderByUsername(String username);
 }
